Reorganize conversion code, remove some duplication
