+++ tezos/api/src/ocaml_conv/mod.rs
@@ -1,0 +1,19 @@
+// Copyright (c) SimpleStaking and Tezedge Contributors
+// SPDX-License-Identifier: MIT
+
+use tezos_messages::p2p::encoding::operations_for_blocks::Path;
+
+// FFI Wrappers
+#[repr(transparent)]
+pub struct FfiPath(pub Path);
+
+// Hashes
+struct OCamlHash {}
+struct OCamlOperationListListHash {}
+pub struct OCamlOperationHash {}
+struct OCamlBlockHash {}
+struct OCamlContextHash {}
+struct OCamlProtocolHash {}
+
+pub mod from_ocaml;
+pub mod to_ocaml;
